* wuwei/dev:
  feat: comment list done v3
  docs: comment list done docs2
  docs: comment list done docs
  feat: comment list done
  feat: comment action done
  feat: favorite list done v2
  feat: favorite list done
  feat: favorite action done 拿掉req的token
  feat: favorite action done
  feat: mysql create 2
  feat: mysql create
  feat:FavoriteAction http done
  feat:gin rpc test done
  feat:test
  feat:contact rpc and gin
  feat:add NewRpcServer to main

# Conflicts:
#	.gitignore
#	idl/interaction.thrift
#	kitex_gen/interaction/interaction.go
#	kitex_gen/interaction/k-interaction.go

+++ models/feed.go
@@ -19,16 +19,16 @@ import (
 //
 //	@Description: 视频数据模型
 type Video struct {
-	ID            uint      `gorm:"primarykey"`
+	ID            int64     `gorm:"primarykey"`
 	CreatedAt     time.Time `gorm:"not null;index:idx_create" json:"created_at,omitempty"`
 	UpdatedAt     time.Time
 	DeletedAt     gorm.DeletedAt `gorm:"index"`
 	Author        User           `gorm:"foreignkey:AuthorID" json:"author,omitempty"`
-	AuthorID      uint           `gorm:"index:idx_authorid;not null" json:"author_id,omitempty"`
+	AuthorID      int64          `gorm:"index:idx_authorid;not null" json:"author_id,omitempty"`
 	PlayUrl       string         `gorm:"type:varchar(512);not null" json:"play_url,omitempty"`
 	CoverUrl      string         `gorm:"type:varchar(255)" json:"cover_url,omitempty"`
-	FavoriteCount uint           `gorm:"default:0;not null" json:"favorite_count,omitempty"`
-	CommentCount  uint           `gorm:"default:0;not null" json:"comment_count,omitempty"`
+	FavoriteCount int64          `gorm:"default:0;not null" json:"favorite_count,omitempty"`
+	CommentCount  int64          `gorm:"default:0;not null" json:"comment_count,omitempty"`
 	Title         string         `gorm:"type:varchar(50);not null" json:"title,omitempty"`
 }
